Use pyproject.toml
setuptools has [deprecated] the test command for the past five years and
recently removed it. This caused [quite a stir] since many people's
build broke.

Our build of [pywb] failed because surt is still using setuptools.test.
I thought perhaps this would be a good time to removing that dependency
and given adoption of [PEP-621] perhaps now is a good time to switch
over to using a `pyproject.toml` file?

[deprecated]: https://github.com/pypa/setuptools/issues/1684
[quite a stir]: https://github.com/pypa/setuptools/issues/4519
[pywb]: https://github.com/webrecorder/pywb
[PEP-621]: https://peps.python.org/pep-0621/